Added more /cube handling code; split edr functionality into several files
